0.1.2: Patient (migration), User (model & logic) - Implemented Patient migration (database table) - Updated User migration (table columns)   - All id's replaced w/ uuid   - Added 'active' columns- no more deleting, only deactivating     - On deactivation, email (unique) overwritten w/ uuid       - Allows creating new accounts w/ prev. used email addresses   - Added 'credentials' column for clinical info

+++ base/resources/views/users/index.blade.php
@@ -30,17 +30,19 @@
                         <tbody>
                         @if (count($users) > 0)
                             @foreach ($users as $user)
-                                <tr>
-                                    <td class="align-content-center text-sm">{{ $user->name }}</td>
-                                    <td class="align-content-center text-sm">{{ $user->email }}</td>
-                                    <td class="align-content-center text-sm">{{ ucfirst($user->role) }}</td>
-                                    <td class="align-content-center text-sm">
-                                        <a href="/users/edit/{{ $user->id }}" class="btn btn-outline-primary px-3 py-1 text-sm">Edit</a>
-                                        @if ($user->id != Auth::user()->id)
-                                            <a href="/users/delete/confirm/{{ $user->id }}" class="btn btn-outline-danger px-3 py-1 text-sm">Delete</a>
-                                        @endif
-                                    </td>
-                                </tr>
+                                @if ($user->active == true)
+                                    <tr>
+                                        <td class="align-content-center text-sm">{{ $user->name }}</td>
+                                        <td class="align-content-center text-sm">{{ $user->email }}</td>
+                                        <td class="align-content-center text-sm">{{ ucfirst($user->role) }}</td>
+                                        <td class="align-content-center text-sm">
+                                            <a href="/users/edit/{{ $user->id }}" class="btn btn-outline-primary px-3 py-1 text-sm">Edit</a>
+                                            @if ($user->id != Auth::user()->id)
+                                                <a href="/users/delete/confirm/{{ $user->id }}" class="btn btn-outline-danger px-3 py-1 text-sm">Delete</a>
+                                            @endif
+                                        </td>
+                                    </tr>
+                                @endif
                             @endforeach
                         @else
                             <tr>
